+++ src/components/10-atoms/text/react/demo.jsx
@@ -3,6 +3,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import changelog from '../CHANGELOG.md';
 import readme from '../README.md';
+import AXAText from './AXATextReact';
 import DemoTextReact from './DemoTextReact';
 import DemoTextReactVersioned from './DemoTextReactVersioned';
 
@@ -11,7 +12,43 @@ export default {
   parameters: {
     readme,
     changelog,
+    controls: { disabled: false },
+    options: { showPanel: true },
   },
+};
+
+const variantOptions = {
+  default: '',
+  'size-1': 'size-1',
+  'size-2': 'size-2',
+  'size-3': 'size-3',
+  bold: 'bold',
+};
+
+export const Story = ({ variant, addSpanTag, content }) => {
+  const div = document.createElement('div');
+  ReactDOM.render(
+    <AXAText variant={variant}>
+      {addSpanTag ? <span>{content}</span> : content}
+    </AXAText>,
+    div
+  );
+  return div;
+};
+
+Story.args = {
+  variant: '',
+  content: `Is your car your pride and joy, or just a means of getting from A to
+  B ? Whichever applies to you, it will certainly have the best
+  insurance with us. Calculate your premium online – You keep your
+  advisor even when you purchase from us online.`,
+  addSpanTag: false,
+};
+
+Story.argTypes = {
+  variant: { control: { type: 'select', options: variantOptions } },
+  addSpanTag: { name: 'add a span tag around the content' },
+  content: { name: 'set content' },
 };
 
 export const DynamicChildrenUnderReact = () => {
@@ -19,6 +56,7 @@ export const DynamicChildrenUnderReact = () => {
   ReactDOM.render(<DemoTextReact />, div);
   return div;
 };
+
 export const CustomVersionedAxatextUnderReact = () => {
   const div = document.createElement('div');
   ReactDOM.render(<DemoTextReactVersioned />, div);
